Update source collection titles/sort if preallocated
If a book was allocated because of versification, it may not have a title or sort number (due to lacking the localized text that would need to be in the RC).

If an update then adds this book to the RC, the collection data should be updated to reflect this.