Update jooq to 3.14.0
Updates jooq to 3.14.0, which supports Kotlin code gen,

As a result, types that were previously potentially null due to unannotated java types are nullable types. Asserting not null is done for now in order to keep consistent with the previous code.

DSL connections are not all implementing AutoCloseable (and instead, connections requiring closing are ClosableDSLConnection) thus the close/using functions were removed in jooq (and as they are no longer needed, they are removed from our code in this PR).