Verse Bridge Support (#767)
* Get marker labels to marker plugin

Sends a list of verse labels to the marker plugin

* WIP

* Audio File Decorator

* Use OratureAudioFile instead of AudioFile

* Don't use a nullable end

* Place markers from markerLabel list

* wip

* Add markers to decorator

* Wip

* Update Source Audio Accessor to open source audio with bridges

Modifies the SourceAudioAccessor to open source audio with verse bridges

* Marker size should be the size of the number of labels provided

As we are now placing labels from a list, the total number of markers should be the size of the list of labels.

* Replace SourceAudioFile with OratureAudioFile

* refactor marker api (#766)

* Wrap markers in a class

* Update OtterAudioFile.kt

* Move away from using cues directly

* Tests pass

* Fix some tests

* Reorganize files

* Check magic number for zip files if extensions are missing

Checks the magic number for the file to see if it is a zip file as a last chance attempt at getting an appropriate file reader before throwing an exception.

* Update OtterAudioFile.kt

* Move trivial addCue implementation to interface

* Add toString to markers

* Update Markers.kt

* Update OratureAudioFile.kt

* Update OratureAudioFile.kt

* Update ConcatenateAudio.kt

* Access end position by the next marker index not assuming the next number

Previously, finding the next marker was searched by the starting position. Instead, look for the next marker.

The problem before was not that this code was incorrect per-say, but old source audio may omit markers rather than have verse bridges.

* convert single digits to bridges

Converts single digit markers to verse bridges.

If a single digit marker exists within the bridge, but not the start, it is dropped (as it was bridged over)

* Marker Parsers and Parsing Results should be internal

As the OratureMarkerParser is meant as the public API, these classes are marked internal.

* Update documentation of WavMetdata

* Update OratureAudioFile.kt

* Delete SourceAudioFile.kt

* Update VerseMarkerModel.kt

* Update Markers.kt

* Use reified types for getting markers

Uses reified types for getting markers by a given marker type.

These types are mapped to enums to preserve some private access modifiers as reified types require the code within them to be public.