Displays existing audio without offset
Defaults to test data, when no existing data is present.

Does not currently offset the starting position of the existing audio in edge cases where more existing audio is retrieved than needed.

I.e. if 4 seconds of existing audio is needed, but 10 is retrieved, it will show seconds 0-4 instead of seconds 6-10