Allocate total verses, keep track of "placed" markers
Rather than adding verses, all markers should be allocated (due to needing the labels to be sourced elsewhere).

As a result, we need to keep track of whether markers are "placed" or not.

Active verses can just be total verses filtered on whether they are placed.

Synchronization needed as a result of multiple thread access.