Use VerseMarker as a public API, VerseNode internal to ChapterRepresentation (#805)
* Move event subscriptions to top level viewmodel

Much of the major narration events for the entire page were handled by the AudioWorkspace, which, should simply be responsible for the visualization and marker interactions. Many such events effect the page as a whole, and thus, need to have data propagated downward to the other ViewModels.

* Allocate total verses, keep track of "placed" markers

Rather than adding verses, all markers should be allocated (due to needing the labels to be sourced elsewhere).
As a result, we need to keep track of whether markers are "placed" or not.
Active verses can just be total verses filtered on whether they are placed.
Synchronization needed as a result of multiple thread access.