VerseNode Ranges (#817)
* Replace end-start calculation with a function

* Refactor Verse Node to contain a range of sectors, not just a start and end

As VerseMarkers can be moved around, bad sectors can begin to arise in the Scratch Audio File and Verses can span across these bad sectors. Rather than one continuous range, this then needs to be accounted for by a list of sectors.

* Introduce ReaderProvider

Rather than get an AudioFileReader from a file, this allows getting the reader other ways, which will be necessary to allow for loading ChapterRepresentation connections into the Audio system

* Update Jackson

Updates the jackson library in order to allow for serialization of IntRange

Adds org.json as a dependency due to the github bug report class- presumably this was previously provided via jackson but now needs it explicitly?

* Fix VerseMarkerAction

Fixes the verse marker action to either take or give frames to adjacent verse nodes.