Fix: seek after re-record (#878)
* Initial draft fix

Still need to seek to the "correct" position on resumeRecordingAgain and iron out some edge cases, but this should keep the markers from disappearing and keep the visualization more consistent.

* Finish the current verse before beginning record-again (#877)

Before proceeding with Record Again, we should first "close" the paused verse by moving its state to "record-again" and enabling the next verse so that the disable and restore functionality will work when we actually perform Record Again.

* Seeks to proper location when resuming re-record

* Initial draft fix

Still need to seek to the "correct" position on resumeRecordingAgain and iron out some edge cases, but this should keep the markers from disappearing and keep the visualization more consistent.

* Seeks to proper location when resuming re-record

* Merge branch 'dh-narration-seek-after-re-record-fix' of https://github.com/Bible-Translation-Tools/Orature into dh-narration-seek-after-re-record-fix

* Removed unnecessary comments

* Fixed typo and added comment

Co-Authored-By: Joe <jsarabia1247@gmail.com>