Removed direct call to bounceAudio and closes the threadpool when necessary
Replacing the direct call to bounceAudio removed the bug where a user could create multiple takes of a chapter by saving a re-recording before the initial chapter take was finished being created.

Calling takeAudioModifier?.close() prevents the pool from remaining open when the program is stopped or when the user navigates away from the Narration page.