Fix for incorrect teleprompter state after navigating between chapters
Marker for chapter X has a label of X and a formatted label of chapter-marker-1. marker for verse X has a label of X and a formatted label of vm-1. This is why for chapter 1, verse 1 is wrong, for chapter 3, verse 3 is wrong

For chapter X, verse X has an incorrect teleprompter state.