Prevents data loss after opening chapter in Ocenaudio (OT-842) (#1050)
* Bounces audio on every new verse once a take file exists

* Fixes incorrect totalVerses size after opening OcenAudio

Fixes bug where opening ocenAudio without all verses being placed caused an error when the user tried to record.

This bug was caused because totalVerses was being overwritten with a list contained only the verses placed at the time the user went to edit the chapter.

* Allows chapter edit actions if any amount of recording exists

* Added modifyAudioDataTask

* Merge branch 'dev' into dh-OT-842-verse-markers-lost-after-opening-ocenaudio

* Allows edit chapter action before recording all verses

Allows the user to edit the chapter before recording all verses.

This is done by creating a "temporary" take each time the user request to edit the chapter before the chapter is complete.

Once the edit chapter action completes, the temporary take is deleted.

* Deletes audio files for temporary takes

Since the wav file produced by creating a temporary take will not be used in the future, we delete the file as soon as possible to avoid storing unnecessary data.

* Removed unnecessary modifyAudioData call

This is no longer necessary since we are creating temporary takes when performing edit chapter actions without all verse markers being placed/recorded

* clean up

* Removed useTemporaryTake property