Add metadata to make burrito export valid
Changed Derived to Source: I incorrectly assumed that Source referred to a Source text, however this is meant as the mutable original (not software modified) burrito. Derived requires a list of operations performed which modified it.

Added a name to Language

Added a currentScope to the AudioFlavor

Added values to Identification

Added an object to Primary for auth