Makes sure that the source.creator does not change
This fixes the bug because updateSource updated the creator from Door43 to Orature, but kept the original path. So later, when we try to generate the path for the source manifest, we are unable to do so, because it is still using the old path generated using Door43 as the source creator.