Non-sequential sectors fix (#1175)
* Accounts for inclusive ends

Similar issue to trying to use arr[arr.size] to index the last element in an array, the actual last element is at arr[arr.size - 1].  So start = scratchAudio.totalFrames is correct if I want to access one past the last sector of audio.

* Corrects the end value of a sector.

Before, we would have verseAudio.totalFrames = 2. The sectors for this should be [0 .. 3], since 2 frames 4 bytes, however, it was [0 .. 4].

Adding 1 after converting the frames to bytes corrects this issue, and removes "dead" space between sectors.

* Ensures no dead space on new verse and verse edit actions

* Updated chapter edit and chapter import fix

Based on meeting with Joe. Uses until (-1) instead of (+1), because using (+1) does not correctly account for varying frame sizes

* Fixed incorrect start for record again action

* Updated return value in finalizeVerse, since we are not finalizing sectors with until

* Updated tests