Fix importing the right sections
Adds filtering out content not matching a particular chapter of u23003 so that those sections can be omitted from reading (needed as orature markers only mark the starting location and don't save the end location).

Fixes an issue with Source Audio where a media format will be selected because of a template entry in the media manifest, and not trying others even if the matching file is not there. This will continue to match other media files until one is found or the search is exhausted.